Refactor: Move `instruction.tokenizer` package from `core` to `lib`, updating imports accordingly
