+++ code/core/src/main/java/org/betonquest/betonquest/api/instruction/DefaultInstruction.java
@@ -14,10 +14,6 @@ import org.betonquest.betonquest.api.instruction.chain.ChainableInstruction;
 import org.betonquest.betonquest.api.instruction.chain.DecoratableChainRetriever;
 import org.betonquest.betonquest.api.instruction.chain.InstructionChainParser;
 import org.betonquest.betonquest.api.instruction.chain.NumberChainRetriever;
-import org.betonquest.betonquest.api.instruction.tokenizer.Tokenizer;
-import org.betonquest.betonquest.api.instruction.tokenizer.TokenizerException;
-import org.betonquest.betonquest.api.instruction.tokenizer.quoting.QuotingTokenizer;
-import org.betonquest.betonquest.api.instruction.tokenizer.quoting.QuotingTokenizerSettings;
 import org.betonquest.betonquest.api.instruction.type.BlockSelector;
 import org.betonquest.betonquest.api.instruction.type.ItemWrapper;
 import org.betonquest.betonquest.api.service.placeholder.PlaceholderManager;
@@ -27,6 +23,10 @@ import org.betonquest.betonquest.lib.instruction.chain.DefaultDecoratableChainRe
 import org.betonquest.betonquest.lib.instruction.chain.DefaultNumberChainRetriever;
 import org.betonquest.betonquest.lib.instruction.reader.DefaultInstructionReader;
 import org.betonquest.betonquest.lib.instruction.reader.SingleValueReader;
+import org.betonquest.betonquest.lib.instruction.tokenizer.Tokenizer;
+import org.betonquest.betonquest.lib.instruction.tokenizer.TokenizerException;
+import org.betonquest.betonquest.lib.instruction.tokenizer.quoting.QuotingTokenizer;
+import org.betonquest.betonquest.lib.instruction.tokenizer.quoting.QuotingTokenizerSettings;
 import org.bukkit.Location;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
