--- code/core/src/main/java/org/betonquest/betonquest/api/instruction/tokenizer/Tokenizer.java
@@ -1,17 +1,0 @@
-package org.betonquest.betonquest.api.instruction.tokenizer;
-
-/**
- * Tokenize an instruction into its parts.
- */
-@FunctionalInterface
-public interface Tokenizer {
-
-    /**
-     * Get the tokens for a given instruction string.
-     *
-     * @param instruction instruction to tokenize
-     * @return tokens of the instruction
-     * @throws TokenizerException if the instruction cannot be tokenized
-     */
-    Token[] tokens(String instruction) throws TokenizerException;
-}
