--- code/core/src/main/java/org/betonquest/betonquest/api/instruction/tokenizer/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Instruction tokenizer related classes.
- */
-package org.betonquest.betonquest.api.instruction.tokenizer;
