--- code/core/src/main/java/org/betonquest/betonquest/api/instruction/tokenizer/quoting/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Classes related to the quoting tokenizer.
- */
-package org.betonquest.betonquest.api.instruction.tokenizer.quoting;
