--- code/lib/src/main/java/org/betonquest/betonquest/lib/instruction/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains default implementations related to instructions.
- */
-package org.betonquest.betonquest.lib.instruction;
