+++ code/lib/src/main/java/org/betonquest/betonquest/lib/instruction/tokenizer/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains default implementations related to instructions.
+ */
+package org.betonquest.betonquest.lib.instruction.tokenizer;
