+++ code/lib/src/main/java/org/betonquest/betonquest/lib/instruction/tokenizer/placeholder/PlaceholderExtractor.java
@@ -1,0 +1,23 @@
+package org.betonquest.betonquest.lib.instruction.tokenizer.placeholder;
+
+import org.betonquest.betonquest.lib.instruction.tokenizer.Token;
+import org.betonquest.betonquest.lib.instruction.tokenizer.Tokenizer;
+import org.betonquest.betonquest.lib.instruction.tokenizer.TokenizerException;
+import org.betonquest.betonquest.lib.instruction.tokenizer.TokenizerState;
+
+/**
+ * Extracts placeholders from an instruction string.
+ */
+public class PlaceholderExtractor implements Tokenizer {
+
+    @Override
+    public Token[] tokens(final String instruction) throws TokenizerException {
+        final PlaceholderExtractorContext ctx = new PlaceholderExtractorContext();
+        TokenizerState<PlaceholderExtractorContext> state = new NoWordState();
+        for (final int character : instruction.codePoints().toArray()) {
+            state = state.parseNext(ctx, character);
+        }
+        state.parseEnd(ctx);
+        return ctx.getTokens().toArray(Token[]::new);
+    }
+}
