+++ code/lib/src/main/java/org/betonquest/betonquest/lib/instruction/tokenizer/placeholder/PlaceholderExtractorSettings.java
@@ -1,0 +1,18 @@
+package org.betonquest.betonquest.lib.instruction.tokenizer.placeholder;
+
+/**
+ * Settings for the placeholder extractor.
+ *
+ * @param placeholderBrackets the brackets that surround the placeholder
+ * @param escapeCharacter     the escape character
+ * @param openQuote           the opening quote character
+ * @param closeQuote          the closing quote character
+ */
+public record PlaceholderExtractorSettings(char placeholderBrackets, char escapeCharacter, char openQuote,
+                                           char closeQuote) {
+
+    /**
+     * The default settings for the placeholder extractor.
+     */
+    public static final PlaceholderExtractorSettings DEFAULT = new PlaceholderExtractorSettings('%', '\\', '{', '}');
+}
