+++ code/lib/src/main/java/org/betonquest/betonquest/lib/instruction/tokenizer/quoting/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes related to the quoting tokenizer.
+ */
+package org.betonquest.betonquest.lib.instruction.tokenizer.quoting;
